archive/legacy/CLOSET_dsa110-contimg/refpipes/ska-ost-low-uv-main/pyproject.toml
[project]
name = "ska-ost-low-uv"
version = "0.3.0"
description = "SKA-low Single-station UV utilities."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "astropy>=6.1.0",
    "cython>=3.1.2",
    "dask>=2025.7.0",
    "h5py>=3.12.0",
    "healpy>=1.17.0",
    "joblib>=1.5.1",
    "loguru>=0.7.3",
    "numpy>=2.1.0",
    "pandas>=2.3.1",
    "pygdsm>=1.6.2",
    "pyuvdata==3.1.3",
    "pyyaml>=6.0.2",
    "tqdm>=4.67.1",
    "xarray>=2024.7.0"
]

[project.optional-dependencies]
postx = [
    "pyradiosky>=1.1.0",
    "sgp4>=2.25",
    "hdf5plugin",
    "makefun>=1.14",
    "jax>=0.7.0"
]
acacia = [
    "rclone-python>=0.1.23",
]
docs = [
    "myst-nb>=1.3.0",
    "ska-ser-sphinx-theme>=0.2.1",
    "sphinx>=8.2.3",
    "sphinx-autodoc-typehints>=3.2.0",
]
test = [
    "colored>=2.3.1",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pytest-mpl>=0.17.0",
]
casa = [
    "python-casacore==3.5.2"
]


[project.scripts]
aa_uv = "ska_ost_low_uv.converter:run"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
ska_ost_low_uv = [
    "config/*.yaml",
    "config/*.txt",
    "config/*.md",
    "config/*/*.yaml",
    "config/*/*.txt",
    "datamodel/*.yaml",
    "postx/data/aeps/*.h5",
    "tests/test-data/*/*.hdf5",
    "tests/test-data/*/*.uv",
    "tests/test-data/*/*.uvfits",
    "tests/test-data/*/*.md",
    "tests/test-data/*/*.uvx",
    "tests/test-data/*/*.txt",
    "tests/test-data/*/*.yaml",
    "tests/test-data/context.yaml"
    ]

[tool.pixi.project]
channels = ["conda-forge"]
platforms = ["osx-arm64"]

[tool.pixi.pypi-dependencies]
ska-ost-low-uv = { path = ".", editable = true }

[tool.pixi.tasks]

[tool.pixi.dependencies]
casacore = "3.5.2"
python-casacore = ">=3.7.1,<4"

[tool.uv.sources]
ska-ser-sphinx-theme = { index = "ska-nexus" }

[[tool.uv.index]]
name = "ska-nexus"
url = "https://artefact.skao.int/repository/pypi-internal/simple"
explicit = true

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
    "uv>=0.8.8",
    "ruff>=0.12.8",
]

# Run with: ruff check --fix
[tool.ruff.lint]
# Setup linting
# E: pycodestyle
# F: Pyflakes
# I: isort
# B: flake8-bugbear
# D: pydocstyle
select = ["E", "F", "I", "B"]
extend-select = ["D"]
ignore = ["E221", "E501", "E741", "D102", "B026"]
# E501 - line too long
# E221 - lining up equals signs
# E741 - ambiguous variable names (using l,m,n)
# D102 - missing docstring in public method; ignore due to docstring inheriting approach
# B026 - Star-arg unpacking after a keyword argument (used to pass *args and **kwargs for plotting)

[tool.ruff.lint.pydocstyle]
convention = "google"

# Run with: ruff format --respect-gitignore --preview --line-length=112
[tool.ruff.format]
quote-style = "single"
docstring-code-line-length = 112

[tool.pytest.ini_options]
addopts = "--cov=ska_ost_low_uv --cov-report html --cov-report xml --mpl --mpl-baseline-path=tests/baseline"
# To regenerate test images, add this argument:
# --mpl-generate-path=tests/baseline
testpaths = ["tests", "notebooks"]

[tool.coverage.run]
branch = true
source = ["ska_ost_low_uv"]
